Some changes
Still compiling with errors, but made some progress by adding more logical progression changes. For properly used props in the Login.js. Added a way to display the email and password submitted. Added a way to check the info coming back from the LoginDatabase.js through an if/else to see if there is a match.