Fixed issues, added new page
Been a while, but here we go.
Fixed issue with project not compiling: missing a / after .. in the import for Login.js. LoginDatabase.js for loop changed from in to of and now properly checks each object in the array instead of being undefined. Added a shop page and will work in the future on how to make the page private unless a proper login is accepted. Shaking the rust off one step at a time.