Added Cart, also many changes
Added the Cart component. This will eventually be moved into a navbar. Lots of styling will be needed for the future as well. Currently need a way to redirect from the login page to the shop page so that it will remember the currently logged in user. Added new databases to keep track of the logged in user and the cart items.