Changes summary below
Added components: AppNavbar, CartMax. Changed Cart component to CartMini; will work on implementing this the way I want to in the future. Added page: CartPage. Began trying to add some styles, grids, and other aspects to the project. Need to properly implement the cart database. Needs plenty more work, but it's a step in the right direction.